+++ R/backend_api-11.ClusterFutureBackend-class.R
@@ -550,7 +550,7 @@ resolved.ClusterFuture <- function(x, run = TRUE, timeout = NULL, ...) {
         )
         avail[nodes] <- FALSE
       }
-      if (debug) mdebugf("avail: [n=%d] %s", length(avail), which(avail))
+      if (debug) mdebugf("avail: [n=%d] %s", length(avail), commaq(which(avail)))
 
       ## Sanity check
       stop_if_not(any(avail))
@@ -584,6 +584,7 @@ resolved.ClusterFuture <- function(x, run = TRUE, timeout = NULL, ...) {
     ## AD HOC/SPECIAL CASE: Skip if connection has been serialized and lacks
     ## internal representation. /HB 2018-10-27
     connId <- connectionId(con)
+    if (debug) mdebugf("Cluster node socket connection: index=%d, id=%d", con, connId)
     if (!is.na(connId) && connId < 0L) return(FALSE)
 
     ## Broken connection due to interruption?
@@ -616,8 +617,11 @@ resolved.ClusterFuture <- function(x, run = TRUE, timeout = NULL, ...) {
     count <- 0L
     while (count < maxCount) {
       ## Is there a message from the worker waiting?
-      res <- socketSelect(list(con), write = FALSE, timeout = timeout)
-      if (!res) break
+      res <- socketSelect(list(con), timeout = timeout, write = FALSE)
+      if (!res) {
+        if (debug) mdebugf("socketSelect(list(<connection #%d (id=%d)>), timeout = %g, write = FALSE) returned %s; not resolved", con, connId, timeout, res)
+        break
+      }
 
       ## Receive it
       msg <- receiveMessageFromWorker(future, debug = debug)
@@ -625,13 +629,29 @@ resolved.ClusterFuture <- function(x, run = TRUE, timeout = NULL, ...) {
       ## If the message contains a FutureResult, then the future is resolved
       ## and we are done here
       res <- inherits(msg, "FutureResult")
-      if (res) break
+      if (res) {
+        if (debug) mdebugf("receiveMessageFromWorker() returned object of class %s; resolved", class(msg)[1])
+        break
+      }
 
       ## If the message contains a FutureInterruptError, then the future
-      ## was interrupted and we are done here
+      ## was interrupted and we are done here. Consider future resolved,
+      ## an let relaying of errors take care of this later.
       res <- inherits(msg, "FutureInterruptError")
-      if (res) break
-
+      if (res) {
+        if (debug) mdebugf("receiveMessageFromWorker() returned object of class %s; resolved", class(msg)[1])
+        break
+      }
+      
+      ## If the message contains a FutureError, e.g. FutureLaunchError, then
+      ## there's nothing more we can do here. Consider future resolved,
+      ## an let relaying of errors take care of this later.
+      res <- inherits(msg, "FutureError")
+      if (res) {
+        if (debug) mdebugf("receiveMessageFromWorker() returned object of class %s; non-fixable state", class(msg)[1])
+        break
+      }
+        
       msg <- NULL
 
       ## If not, we received a condition that has already been signaled
@@ -649,11 +669,13 @@ resolved.ClusterFuture <- function(x, run = TRUE, timeout = NULL, ...) {
 
   if (debug) mdebug_pop()
 
-  ## Assert result is for the expected future
-  assertFutureResult(future)
+  if (res) {
+    ## Assert result is for the expected future
+    assertFutureResult(future, debug = debug)
 
-  ## Signal conditions early? (happens only iff requested)
-  if (res) signalEarly(future, ...)
+    ## Signal conditions early? (happens only iff requested)
+    signalEarly(future, ...)
+  }
 
   res
 }
@@ -803,8 +825,6 @@ receiveMessageFromWorker <- local({
         stop(ex)
       }
 
-      str(list(data = data, msg = msg))
-      
       node_info <- sprintf("%s #%d", sQuote(class(node)[1]), node_idx)
       if (inherits(node, "RichSOCKnode")) {
         specs <- summary(node)
@@ -884,6 +904,9 @@ receiveMessageFromWorker <- local({
         if (debug) mdebug_pop()
       }
     } else if (inherits(msg, "FutureLaunchError")) {
+      if (debug) mdebugf("Received %s", class(msg)[1])
+      future[["result"]] <- msg
+      future[["state"]] <- "failed"
     } else if (inherits(msg, "condition")) {
       condition <- msg
       
@@ -894,6 +917,8 @@ receiveMessageFromWorker <- local({
   
       ## Sanity check
       if (inherits(condition, "error")) {
+        future[["result"]] <- msg
+        future[["state"]] <- "failed"
         label <- sQuoteLabel(future[["label"]])
         stop(FutureError(sprintf("Received a %s condition from the %s worker for future (%s), which is not possible to relay because that would break the internal state of the future-worker communication. The condition message was: %s", class(condition)[1], class(future)[1], label, sQuote(conditionMessage(condition))), future = future))
       }
@@ -1013,7 +1038,7 @@ requestNode <- function(await, backend, timeout, delta, alpha) {
   ## Sanity check
   stop_if_not(any(avail))
 
-  if (debug) mdebugf("avail: [n=%d] %s", length(avail), which(avail))
+  if (debug) mdebugf("avail: [n=%d] %s", length(avail), commaq(which(avail)))
 
   node_idx <- which(avail)[1L]
   stop_if_not(is.numeric(node_idx), is.finite(node_idx), node_idx >= 1, node_idx <= length(workers))